feat: migrate channel flow analytics methods to forwarding HTLC event repository and remove unused repository interfaces
