--- src/Data/Repositories/Interfaces/IChannelFlowAnalyticsRepository.cs
@@ -1,46 +1,0 @@
-/*
- * NodeGuard
- * Copyright (C) 2023  Elenpay
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see http://www.gnu.org/licenses/.
- *
- */
-
-namespace NodeGuard.Data.Repositories.Interfaces;
-
-/// <summary>
-/// Windowed per-channel aggregations over ForwardingHtlcEvent. Succeeded HTLCs only — every
-/// query filters Outcome == Settled, so failed and in-flight events never count as flow.
-/// </summary>
-public interface IChannelFlowAnalyticsRepository
-{
-    /// <summary>
-    /// Σ OutgoingAmountMsat of succeeded forwards where OutgoingChannelId == chanIdLnd, for the
-    /// given managed node, since the window start (drains our local balance — "push").
-    /// </summary>
-    Task<long> GetOutgoingAmountMsat(string managedNodePubKey, ulong chanIdLnd, DateTimeOffset since);
-
-    /// <summary>
-    /// Σ IncomingAmountMsat of succeeded forwards where IncomingChannelId == chanIdLnd, for the
-    /// given managed node, since the window start (fills our local balance — "pull").
-    /// </summary>
-    Task<long> GetIncomingAmountMsat(string managedNodePubKey, ulong chanIdLnd, DateTimeOffset since);
-
-    /// <summary>
-    /// Σ FeeMsat of succeeded forwards where OutgoingChannelId == chanIdLnd, since the window
-    /// start. Attributed to the outgoing side (the channel whose liquidity was spent) so
-    /// per-channel revenue sums don't double-count. Phase 2 prioritization input.
-    /// </summary>
-    Task<long> GetOrganicFeesEarnedMsat(string managedNodePubKey, ulong chanIdLnd, DateTimeOffset since);
-}
